// Web/RMA-Member-Portal/memberportal-app/src/app/shared/enums/role-player-type-enum.ts
export enum RolePlayerTypeEnum {
  PolicyOwner = 1,
  PolicyPayee = 2,
  InsuredLife = 3,
  FinancialServiceProvider = 4,
  Claimant = 5,
  MedicalServiceProvider = 6,
  FuneralParlor = 7,
  BodyCollector = 8,
  Undertaker = 9,
  MainMemberSelf = 10,
  Spouse = 11,
  Daughter = 12,
  DaughterInlaw = 13,
  SonInlaw = 14,
  Son = 15,
  Parent = 16,
  ParentInlaw = 17,
  Grandparent = 18,
  Mother = 19,
  MotherInlaw = 20,
  Father = 21,
  FatherInlaw = 22,
  Brother = 23,
  BrotherInlaw = 24,
  Sister = 25,
  SisterInlaw = 26,
  Aunt = 27,
  Niece = 28,
  Nephew = 29,
  Husband = 30,
  Wife = 31,
  Child = 32,
  SpecialChild = 33,
  Other = 34,
  GuardianRecipient = 35,
  PersonIndividual = 36,
  Pensioner = 37,
  Extended = 38,
  DisabledChild = 39,
  Uncle = 40,
  Beneficiary = 41,
  Cousin = 42,
  GrandChild = 44,
  Friend = 45
}